Create theme test with SSDC-PM docs
